Rename package "varsel" to "planner"
