+++ src/main/kotlin/no/nav/syfo/planner/AktivitetskravVarselPlannerSyketilfellebit.kt
@@ -1,7 +1,6 @@
-package no.nav.syfo.varsel
+package no.nav.syfo.planner
 
 import kotlinx.coroutines.coroutineScope
-import no.nav.syfo.consumer.syfosyketilfelle.SyfosyketilfelleConsumer
 import no.nav.syfo.db.DatabaseInterface
 import no.nav.syfo.db.deletePlanlagtVarselBySykmeldingerId
 import no.nav.syfo.db.domain.PlanlagtVarsel
@@ -9,28 +8,23 @@ import no.nav.syfo.db.domain.VarselType
 import no.nav.syfo.db.storePlanlagtVarsel
 import no.nav.syfo.metrics.tellAktivitetskravPlanlagt
 import no.nav.syfo.service.SykmeldingService
+import no.nav.syfo.syketilfelle.SyketilfellebitService
 import no.nav.syfo.utils.VarselUtil
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
-class AktivitetskravVarselPlanner(
-    private val databaseAccess: DatabaseInterface,
-    private val syfosyketilfelleConsumer: SyfosyketilfelleConsumer,
-    val sykmeldingService: SykmeldingService,
-    override val name: String = "AKTIVITETSKRAV_VARSEL"
-) : VarselPlanner {
-    private val AKTIVITETSKRAV_DAGER: Long = 42
-
-    private val log: Logger = LoggerFactory.getLogger("no.nav.syfo.varsel.AktivitetskravVarselPlanner")
+class AktivitetskravVarselPlannerSyketilfellebit(
+    val databaseAccess: DatabaseInterface,
+    val syketilfellebitService: SyketilfellebitService,
+    val sykmeldingService: SykmeldingService
+) : VarselPlannerSyketilfellebit {
+    private val log: Logger = LoggerFactory.getLogger("no.nav.syfo.varsel.AktivitetskravVarselPlannerSyketilfellebit")
     private val varselUtil: VarselUtil = VarselUtil(databaseAccess)
+    private val AKTIVITETSKRAV_DAGER: Long = 42L
+    override val name: String = "AKTIVITETSKRAV_VARSEL_GCP"
 
-    override suspend fun processOppfolgingstilfelle(aktorId: String, fnr: String, orgnummer: String?) = coroutineScope {
-        val oppfolgingstilfellePerson = syfosyketilfelleConsumer.getOppfolgingstilfelle(aktorId)
-
-        if (oppfolgingstilfellePerson == null) {
-            log.info("-$name-: Fant ikke oppfolgingstilfelle. Planlegger ikke nytt varsel")
-            return@coroutineScope
-        }
+    override suspend fun processSyketilfelle(fnr: String, orgnummer: String) = coroutineScope {
+        val oppfolgingstilfellePerson = syketilfellebitService.beregnKOppfolgingstilfelle(fnr) ?: return@coroutineScope
 
         val validSyketilfelledager = oppfolgingstilfellePerson.tidslinje.filter { varselUtil.isValidSyketilfelledag(it) }.sortedBy { it.dag }
 
@@ -114,7 +108,8 @@ class AktivitetskravVarselPlanner(
                 databaseAccess.storePlanlagtVarsel(aktivitetskravVarsel)
                 tellAktivitetskravPlanlagt()
             }
+        } else {
+            log.info("-$name-: Ingen gyldigeSykmeldingTilfelledager. Planlegger ikke nytt varsel")
         }
-        log.info("-$name-: Ingen gyldigeSykmeldingTilfelledager. Planlegger ikke nytt varsel")
     }
 }
