+++ src/main/kotlin/no/nav/syfo/planner/MerVeiledningVarselPlanner.kt
@@ -1,6 +1,7 @@
-package no.nav.syfo.varsel
+package no.nav.syfo.planner
 
 import kotlinx.coroutines.coroutineScope
+import no.nav.syfo.consumer.syfosyketilfelle.SyfosyketilfelleConsumer
 import no.nav.syfo.db.DatabaseInterface
 import no.nav.syfo.db.deletePlanlagtVarselByVarselId
 import no.nav.syfo.db.domain.PlanlagtVarsel
@@ -9,23 +10,27 @@ import no.nav.syfo.db.storePlanlagtVarsel
 import no.nav.syfo.db.updateUtsendingsdatoByVarselId
 import no.nav.syfo.metrics.tellMerVeiledningPlanlagt
 import no.nav.syfo.service.VarselSendtService
-import no.nav.syfo.syketilfelle.SyketilfellebitService
 import no.nav.syfo.utils.VarselUtil
 import no.nav.syfo.utils.todayIsBetweenFomAndTom
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
-class MerVeiledningVarselPlannerSyketilfellebit(
+class MerVeiledningVarselPlanner(
     val databaseAccess: DatabaseInterface,
-    val syketilfellebitService: SyketilfellebitService,
+    val syfosyketilfelleConsumer: SyfosyketilfelleConsumer,
     val varselSendtService: VarselSendtService
-) : VarselPlannerSyketilfellebit {
-    private val log: Logger = LoggerFactory.getLogger("no.nav.syfo.varsel.MerVeiledningVarselPlannerSyketilfellebit")
+) : VarselPlanner {
+    private val log: Logger = LoggerFactory.getLogger("no.nav.syfo.varsel.Varsel39Uker")
     private val varselUtil: VarselUtil = VarselUtil(databaseAccess)
-    override val name: String = "MER_VEILEDNING_VARSEL_GCP"
+    override val name: String = "MER_VEILEDNING_VARSEL"
 
-    override suspend fun processSyketilfelle(fnr: String, orgnummer: String) = coroutineScope {
-        val oppfolgingstilfelle = syketilfellebitService.beregnKOppfolgingstilfelle39UkersVarsel(fnr) ?: return@coroutineScope
+    override suspend fun processOppfolgingstilfelle(aktorId: String, fnr: String, orgnummer: String?) = coroutineScope {
+        val oppfolgingstilfelle = syfosyketilfelleConsumer.getOppfolgingstilfelle39Uker(aktorId)
+
+        if (oppfolgingstilfelle == null) {
+            log.info("[$name]: Fant ikke oppfolgingstilfelle for denne brukeren. Planlegger ikke nytt varsel")
+            return@coroutineScope
+        }
 
         val tilfelleFom = oppfolgingstilfelle.fom
         val tilfelleTom = oppfolgingstilfelle.tom
