+++ src/main/kotlin/no/nav/syfo/planner/SvarMotebehovVarselPlannerSyketilfellebit.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.varsel
+package no.nav.syfo.planner
 
 import kotlinx.coroutines.coroutineScope
 import no.nav.syfo.db.DatabaseInterface
