+++ src/main/kotlin/no/nav/syfo/planner/VarselPlanner.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.varsel
+package no.nav.syfo.planner
 
 interface VarselPlanner {
     val name: String
