+++ src/test/kotlin/no/nav/syfo/planner/AktivitetskravVarselPlannerSpek.kt
@@ -1,20 +1,20 @@
-package no.nav.syfo.varsel
+package no.nav.syfo.planner
 
 import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.runBlocking
+import no.nav.syfo.consumer.syfosyketilfelle.SyfosyketilfelleConsumer
 import no.nav.syfo.consumer.syfosmregister.SykmeldingerConsumer
+import no.nav.syfo.kafka.consumers.oppfolgingstilfelle.domain.Syketilfelledag
 import no.nav.syfo.consumer.syfosmregister.SykmeldtStatusResponse
+import no.nav.syfo.syketilfelle.domain.Syketilfellebit
 import no.nav.syfo.db.domain.PlanlagtVarsel
 import no.nav.syfo.db.domain.VarselType
 import no.nav.syfo.db.fetchPlanlagtVarselByFnr
 import no.nav.syfo.db.storePlanlagtVarsel
-import no.nav.syfo.kafka.consumers.oppfolgingstilfelle.domain.Syketilfelledag
+import no.nav.syfo.kafka.consumers.oppfolgingstilfelle.domain.OppfolgingstilfellePerson
 import no.nav.syfo.service.SykmeldingService
-import no.nav.syfo.syketilfelle.KOppfolgingstilfellePerson
-import no.nav.syfo.syketilfelle.SyketilfellebitService
-import no.nav.syfo.syketilfelle.domain.Syketilfellebit
-import no.nav.syfo.syketilfelle.domain.Tag
+import no.nav.syfo.syketilfelle.domain.Tag.*
 import no.nav.syfo.testutil.EmbeddedDatabase
 import no.nav.syfo.testutil.dropData
 import no.nav.syfo.utils.VarselUtil
@@ -27,19 +27,19 @@ import java.time.LocalDateTime
 import java.time.temporal.ChronoUnit
 import kotlin.test.assertFailsWith
 
-object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
+object AktivitetskravVarselPlannerSpek : Spek({
     // The default timeout of 10 seconds is not sufficient to initialise the embedded database
     defaultTimeout = 20000L
 
     val embeddedDatabase by lazy { EmbeddedDatabase() }
     val sykmeldingerConsumer = mockk<SykmeldingerConsumer>()
-    val syketilfellebitService = mockk<SyketilfellebitService>()
+    val syfosyketilfelleConsumer = mockk<SyfosyketilfelleConsumer>()
     val varselUtil = VarselUtil(embeddedDatabase)
 
     val aktivitetskravVarselPlanner =
-        AktivitetskravVarselPlannerSyketilfellebit(embeddedDatabase, syketilfellebitService, SykmeldingService(sykmeldingerConsumer))
+        AktivitetskravVarselPlanner(embeddedDatabase, syfosyketilfelleConsumer, SykmeldingService(sykmeldingerConsumer))
 
-    describe("AktivitetskravVarselPlannerSyketilfellebitSpek") {
+    describe("AktivitetskravVarselPlannerSpek") {
         val planlagtVarselToStore2 =
             PlanlagtVarsel(arbeidstakerFnr1, arbeidstakerAktorId1, orgnummer, setOf("1"), VarselType.MER_VEILEDNING)
         val planlagtVarselToStore3 =
@@ -67,7 +67,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                     "2",
                     LocalDateTime.now(),
                     LocalDateTime.now(),
-                    listOf(Tag.ANNET_FRAVAR, Tag.SENDT),
+                    listOf(ANNET_FRAVAR, SENDT),
                     "3",
                     LocalDate.now(),
                     LocalDate.now().plusDays(7)
@@ -79,7 +79,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                     "2",
                     LocalDateTime.now(),
                     LocalDateTime.now(),
-                    listOf(Tag.ANNET_FRAVAR, Tag.SENDT),
+                    listOf(ANNET_FRAVAR, SENDT),
                     "3",
                     LocalDate.now().plusDays(8),
                     LocalDate.now().plusDays(16)
@@ -91,7 +91,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                     "2",
                     LocalDateTime.now(),
                     LocalDateTime.now(),
-                    listOf(Tag.ANNET_FRAVAR, Tag.SENDT),
+                    listOf(ANNET_FRAVAR, SENDT),
                     "3",
                     LocalDate.now().plusDays(17),
                     LocalDate.now().plusDays(50)
@@ -108,7 +108,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                 tom = LocalDate.now().plusDays(50)
             )
             val oppfolgingstilfellePerson =
-                KOppfolgingstilfellePerson(
+                OppfolgingstilfellePerson(
                     arbeidstakerFnr1,
                     listOf(syketilfelledag1, syketilfelledag2, syketilfelledag3),
                     syketilfelledag1,
@@ -118,10 +118,10 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                 )
 
             coEvery { sykmeldingerConsumer.getSykmeldtStatusPaDato(any(), any()) } returns sykmeldtStatusResponse
-            coEvery { syketilfellebitService.beregnKOppfolgingstilfelle(any()) } returns oppfolgingstilfellePerson
+            coEvery { syfosyketilfelleConsumer.getOppfolgingstilfelle(any()) } returns oppfolgingstilfellePerson
 
             runBlocking {
-                aktivitetskravVarselPlanner.processSyketilfelle(arbeidstakerFnr1, orgnummer)
+                aktivitetskravVarselPlanner.processOppfolgingstilfelle(arbeidstakerAktorId1, arbeidstakerFnr1, orgnummer)
 
                 val lagreteVarsler = embeddedDatabase.fetchPlanlagtVarselByFnr(arbeidstakerFnr1)
                 val nrOfRowsFetchedTotal = lagreteVarsler.size
@@ -141,7 +141,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                     arbeidstakerAktorId1,
                     "2", LocalDateTime.now(),
                     LocalDateTime.now(),
-                    listOf(Tag.SYKMELDING, Tag.SENDT),
+                    listOf(SYKMELDING, SENDT),
                     "3",
                     LocalDate.now().minusDays(100),
                     LocalDate.now().minusDays(100)
@@ -149,7 +149,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
 
             val syketilfelledag1 = Syketilfelledag(LocalDate.now().minusDays(60), syketilfellebit1)
             val oppfolgingstilfellePerson =
-                KOppfolgingstilfellePerson(
+                OppfolgingstilfellePerson(
                     arbeidstakerFnr1,
                     listOf(syketilfelledag1),
                     syketilfelledag1,
@@ -162,10 +162,10 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                 SykmeldtStatusResponse(erSykmeldt = true, gradert = false, fom = LocalDate.now(), tom = LocalDate.now())
 
             coEvery { sykmeldingerConsumer.getSykmeldtStatusPaDato(any(), any()) } returns sykmeldtStatusResponse
-            coEvery { syketilfellebitService.beregnKOppfolgingstilfelle(any()) } returns oppfolgingstilfellePerson
+            coEvery { syfosyketilfelleConsumer.getOppfolgingstilfelle(any()) } returns oppfolgingstilfellePerson
 
             runBlocking {
-                aktivitetskravVarselPlanner.processSyketilfelle(arbeidstakerFnr1, orgnummer)
+                aktivitetskravVarselPlanner.processOppfolgingstilfelle(arbeidstakerAktorId1, arbeidstakerFnr1, orgnummer)
 
                 val lagreteVarsler = embeddedDatabase.fetchPlanlagtVarselByFnr(arbeidstakerFnr1)
                 val nrOfRowsFetchedTotal = lagreteVarsler.size
@@ -187,7 +187,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                     "2",
                     LocalDateTime.now(),
                     LocalDateTime.now(),
-                    listOf(Tag.SYKMELDING, Tag.SENDT),
+                    listOf(SYKMELDING, SENDT),
                     "3",
                     LocalDate.now(),
                     LocalDate.now()
@@ -199,7 +199,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                     "2",
                     LocalDateTime.now(),
                     LocalDateTime.now(),
-                    listOf(Tag.SYKMELDING, Tag.SENDT),
+                    listOf(SYKMELDING, SENDT),
                     "3",
                     LocalDate.now(),
                     LocalDate.now()
@@ -211,7 +211,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                     "2",
                     LocalDateTime.now(),
                     LocalDateTime.now(),
-                    listOf(Tag.PAPIRSYKMELDING, Tag.SENDT),
+                    listOf(PAPIRSYKMELDING, SENDT),
                     "3",
                     LocalDate.now(),
                     LocalDate.now()
@@ -224,7 +224,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
             val sykmeldtStatusResponse =
                 SykmeldtStatusResponse(erSykmeldt = true, gradert = false, fom = LocalDate.now(), tom = LocalDate.now())
             val oppfolgingstilfellePerson =
-                KOppfolgingstilfellePerson(
+                OppfolgingstilfellePerson(
                     arbeidstakerFnr1,
                     listOf(syketilfelledag1, syketilfelledag2, syketilfelledag3),
                     syketilfelledag1,
@@ -234,10 +234,11 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                 )
 
             coEvery { sykmeldingerConsumer.getSykmeldtStatusPaDato(any(), any()) } returns sykmeldtStatusResponse
-            coEvery { syketilfellebitService.beregnKOppfolgingstilfelle(any()) } returns oppfolgingstilfellePerson
+            coEvery { syfosyketilfelleConsumer.getOppfolgingstilfelle(any()) } returns oppfolgingstilfellePerson
 
             runBlocking {
-                aktivitetskravVarselPlanner.processSyketilfelle(arbeidstakerFnr1, orgnummer)
+
+                aktivitetskravVarselPlanner.processOppfolgingstilfelle(arbeidstakerAktorId1, arbeidstakerFnr1, orgnummer)
 
                 val lagreteVarsler = embeddedDatabase.fetchPlanlagtVarselByFnr(arbeidstakerFnr1)
                 val nrOfRowsFetchedTotal = lagreteVarsler.size
@@ -259,7 +260,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                     "2",
                     LocalDateTime.now(),
                     LocalDateTime.now(),
-                    listOf(Tag.SYKMELDING, Tag.SENDT),
+                    listOf(SYKMELDING, SENDT),
                     "3",
                     LocalDate.now(),
                     LocalDate.now().plusDays(7)
@@ -271,7 +272,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                     "2",
                     LocalDateTime.now(),
                     LocalDateTime.now(),
-                    listOf(Tag.SYKMELDING, Tag.SENDT),
+                    listOf(SYKMELDING, SENDT),
                     "3",
                     LocalDate.now().plusDays(8),
                     LocalDate.now().plusDays(16)
@@ -283,7 +284,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                     "2",
                     LocalDateTime.now(),
                     LocalDateTime.now(),
-                    listOf(Tag.SYKMELDING, Tag.SENDT),
+                    listOf(SYKMELDING, SENDT),
                     "3",
                     LocalDate.now().plusDays(17),
                     LocalDate.now().plusDays(50)
@@ -300,7 +301,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                 tom = LocalDate.now().plusDays(50)
             )
             val oppfolgingstilfellePerson =
-                KOppfolgingstilfellePerson(
+                OppfolgingstilfellePerson(
                     arbeidstakerFnr1,
                     listOf(syketilfelledag1, syketilfelledag2, syketilfelledag3),
                     syketilfelledag1,
@@ -310,10 +311,10 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                 )
 
             coEvery { sykmeldingerConsumer.getSykmeldtStatusPaDato(any(), any()) } returns sykmeldtStatusResponse
-            coEvery { syketilfellebitService.beregnKOppfolgingstilfelle(any()) } returns oppfolgingstilfellePerson
+            coEvery { syfosyketilfelleConsumer.getOppfolgingstilfelle(any()) } returns oppfolgingstilfellePerson
 
             runBlocking {
-                aktivitetskravVarselPlanner.processSyketilfelle(arbeidstakerFnr1, orgnummer)
+                aktivitetskravVarselPlanner.processOppfolgingstilfelle(arbeidstakerAktorId1, arbeidstakerFnr1, orgnummer)
 
                 val lagreteVarsler = embeddedDatabase.fetchPlanlagtVarselByFnr(arbeidstakerFnr1)
                 val nrOfRowsFetchedTotal = lagreteVarsler.size
@@ -344,7 +345,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
                     "2",
                     LocalDateTime.now(),
                     LocalDateTime.now(),
-                    listOf(Tag.SYKMELDING, Tag.SENDT),
+                    listOf(SYKMELDING, SENDT),
                     "1",
                     utsendingsdato.minusDays(AKTIVITETSKRAV_DAGER).atStartOfDay().toLocalDate(),
                     utsendingsdato.plusDays(30).atStartOfDay().toLocalDate()
@@ -355,7 +356,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
 
             val sykmeldtStatusResponse =
                 SykmeldtStatusResponse(erSykmeldt = true, gradert = false, fom = LocalDate.now(), tom = LocalDate.now())
-            val oppfolgingstilfellePerson = KOppfolgingstilfellePerson(
+            val oppfolgingstilfellePerson = OppfolgingstilfellePerson(
                 arbeidstakerFnr1,
                 listOf(syketilfelledag1, syketilfelledag2),
                 syketilfelledag1,
@@ -365,7 +366,7 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
             )
 
             coEvery { sykmeldingerConsumer.getSykmeldtStatusPaDato(any(), any()) } returns sykmeldtStatusResponse
-            coEvery { syketilfellebitService.beregnKOppfolgingstilfelle(any()) } returns oppfolgingstilfellePerson
+            coEvery { syfosyketilfelleConsumer.getOppfolgingstilfelle(any()) } returns oppfolgingstilfellePerson
 
             val lagreteVarsler1 = embeddedDatabase.fetchPlanlagtVarselByFnr(arbeidstakerFnr1)
             val nrOfRowsFetchedTotal1 = lagreteVarsler1.size
@@ -374,7 +375,8 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
 
             // Skal lagre varsel
             runBlocking {
-                aktivitetskravVarselPlanner.processSyketilfelle(
+                aktivitetskravVarselPlanner.processOppfolgingstilfelle(
+                    arbeidstakerAktorId1,
                     arbeidstakerFnr1,
                     orgnummer
                 )
@@ -388,9 +390,10 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
 
         it("AkvitietskravVarselPlanner kaster RuntimeException dersom syfosyketilfelleConsumer kaster exception") {
             assertFailsWith(RuntimeException::class) {
-                coEvery { syketilfellebitService.beregnKOppfolgingstilfelle(any()) } throws RuntimeException()
+                coEvery { syfosyketilfelleConsumer.getOppfolgingstilfelle(any()) } throws RuntimeException()
                 runBlocking {
-                    aktivitetskravVarselPlanner.processSyketilfelle(
+                    aktivitetskravVarselPlanner.processOppfolgingstilfelle(
+                        arbeidstakerAktorId1,
                         arbeidstakerFnr1,
                         orgnummer
                     )
@@ -400,14 +403,15 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
 
         it("AkvitietskravVarselPlanner dropper planlegging når syfosyketilfelle returnerer null") {
 
-            coEvery { syketilfellebitService.beregnKOppfolgingstilfelle(any()) } returns null
+            coEvery { syfosyketilfelleConsumer.getOppfolgingstilfelle(any()) } returns null
 
             runBlocking {
-                aktivitetskravVarselPlanner.processSyketilfelle(arbeidstakerFnr1, orgnummer)
+                aktivitetskravVarselPlanner.processOppfolgingstilfelle(arbeidstakerAktorId1, arbeidstakerFnr1, orgnummer)
                 val lagreteVarsler = embeddedDatabase.fetchPlanlagtVarselByFnr(arbeidstakerFnr1)
 
                 lagreteVarsler.size shouldBeEqualTo 0
             }
+
         }
 
         it("AkvitietskravVarselPlanner fjerner perioder uten sykmelding når den faktiske fraværlengde blir beregnet") {
@@ -437,3 +441,12 @@ object AktivitetskravVarselPlannerSyketilfellebitSpek : Spek({
         }
     }
 })
+
+fun createValidSyketilfelledag(fom: LocalDateTime, tom: LocalDateTime): Syketilfelledag {
+    return Syketilfelledag(
+        LocalDate.now(),
+        Syketilfellebit(
+            "", "", "", LocalDateTime.now(), LocalDateTime.now(), listOf(SYKMELDING, BEKREFTET), "", fom.toLocalDate(), tom.toLocalDate()
+        )
+    )
+}
