+++ src/test/kotlin/no/nav/syfo/testutil/mocks/MockVarselPlaner.kt
@@ -2,7 +2,7 @@ package no.nav.syfo.testutil.mocks
 
 import kotlinx.coroutines.coroutineScope
 import no.nav.syfo.ApplicationState
-import no.nav.syfo.varsel.VarselPlanner
+import no.nav.syfo.planner.VarselPlanner
 
 class MockVarselPlaner(val applicationState: ApplicationState) : VarselPlanner {
     override val name: String = "MOCK_PLANNER"
